requirements.txt
python-dotenv==1.0.1
google-generativeai==0.7.2

sentence-transformers==3.0.1
transformers==4.45.2
huggingface-hub==0.24.6

protobuf==4.25.3

chromadb==0.5.5

pandas==2.2.2
requests==2.32.3

spacy==3.7.5
textblob==0.18.0

scikit-learn==1.5.1

gradio==4.44.0